Diverse gui-justeringer tilsagnssiden og endring av navngivning
