+++ frontend/mr-admin-flate/src/pages/gjennomforing/tilsagn/AarsakerOgForklaring.tsx
@@ -1,4 +1,4 @@
-import { Alert, BodyShort, Heading, List } from "@navikt/ds-react";
+import { Alert, BodyShort, Heading, HStack, List } from "@navikt/ds-react";
 
 type Props = {
   heading: string;
@@ -11,22 +11,31 @@ type Props = {
 export function AarsakerOgForklaring({ heading, ingress, tekster, aarsaker, forklaring }: Props) {
   return (
     <Alert size="small" variant="warning">
-      <Heading size="small" level="4" className="mb-1">
+      <Heading spacing size="small" level="4">
         {heading}
       </Heading>
-      {ingress && <BodyShort className="mb-6">{ingress}</BodyShort>}
-      <div className="mb-6">
-        {tekster && tekster.map((tekst, index) => <BodyShort key={index}>{tekst}</BodyShort>)}
-      </div>
-      <List>
+      {ingress && <BodyShort className="mb-4">{ingress}</BodyShort>}
+      {tekster &&
+        tekster.map((tekst, index) => (
+          <BodyShort key={index} className="mb-4">
+            {tekst}
+          </BodyShort>
+        ))}
+      <Heading level="5" size="xsmall">
+        Årsaker:
+      </Heading>
+      <List as="ul" size="small">
         {aarsaker.map((aarsak) => (
           <List.Item key={aarsak}>{aarsak}</List.Item>
         ))}
       </List>
       {forklaring && (
-        <BodyShort>
-          <b>Forklaring:</b> {forklaring}
-        </BodyShort>
+        <HStack gap="2">
+          <Heading level="5" size="xsmall">
+            Forklaring:
+          </Heading>
+          <BodyShort>{forklaring}</BodyShort>
+        </HStack>
       )}
     </Alert>
   );
